Bugfix for #21 & #22 -- Add repo constructable from background thread
Common
- add new HeadlessRepository
  - does not expose anything in the form of LiveData, hence safe for Koin to construct in a background context (Workers on Android platform)
- move AnalTrackingRepository and GeoTrackingRepository upload and purge features to HeadlessRepository
- add HeadlessRepository to Koin module
- change target repository in related use case classes

- optimize upload process
  - retrieve some data from key/store local storage once instead of once per record (that was quite alpha)
  - better handle different success and failure conditions to send back as Reponse.Success or Response.Error
