+++ app/src/commonMain/kotlin/com/ludoscity/herdr/common/data/repository/GeoTrackingRepository.kt
@@ -36,16 +36,9 @@ import org.koin.core.parameter.parametersOf
 
 class GeoTrackingRepository : KoinComponent {
 
-    companion object {
-        const val UPLOAD_GEO_PERIODIC_WORKER_UNIQUE_NAME = "herdr-upload-geo-worker"
-        const val PURGE_GEO_PERIODIC_WORKER_UNIQUE_NAME = "herdr-purge-geo-worker"
-    }
-
     private val log: Kermit by inject { parametersOf("GeoTrackingRepository") }
 
     private val herdrDb: HerdrDatabase by inject()
-    private val networkDataPipe: INetworkDataPipe by inject()
-    private val secureDataStore: SecureDataStore by inject()
 
 
     private val _userLoc = MutableLiveData<GeoTrackingDatapoint?>(null)
@@ -83,71 +76,6 @@ class GeoTrackingRepository : KoinComponent {
         //TODO: retrieve timestamp epoch here for both platforms at once
         geoTrackingDao.insert(record)
         return Response.Success(geoTrackingDao.select())
-    }
-
-    suspend fun uploadAllGeoTrackingDatapointReadyForUpload(): Response<Unit> {
-        val geoTrackingDao = GeoTrackingDatapointDao(herdrDb)
-        var atLeastOneError = true
-        geoTrackingDao.selectReadyForUploadAll().forEach {
-            secureDataStore.retrieveString(
-                LoginRepository.authClientRegistrationBaseUrlStoreKey
-            )?.let { stackBase ->
-                log.d { "We have a stack address" }
-                secureDataStore.retrieveString(
-                    LoginRepository.cloudDirectoryId
-                )?.let { cloudDirectoryId ->
-                    log.d { "We have a directory id" }
-                    val networkReply = networkDataPipe.postFile(
-                        stackBase,
-                        cloudDirectoryId,
-                        "${it.timestamp}_GEOLOCATION.json",
-                        listOf("herdr", "geolocation"),
-                        Json(JsonConfiguration.Stable).stringify(
-                            GeoTrackingUploadRequestBody.serializer(),
-                            GeoTrackingUploadRequestBody(
-                                it.timestamp_epoch,
-                                it.altitude,
-                                it.accuracy_horizontal_meters,
-                                it.accuracy_vertical_meters,
-                                it.latitude,
-                                it.longitude,
-                                it.timestamp
-                            )
-                        ),
-                        it.timestamp
-                    )
-
-                    when (networkReply) {
-                        is Response.Success -> {
-                            log.d { "Upload success, flagged record of name: ${it.timestamp}_GEOLOCATION.json for deletion" }
-                            geoTrackingDao.updateUploadCompleted(it.id)
-                            atLeastOneError = false
-                        }
-                        is Response.Error -> {
-                            if (networkReply.code == 409) {
-                                log.i { "Recovered from 409, flagged record: ${it.timestamp}_GEOLOCATION.json for deletion" }
-                                geoTrackingDao.updateUploadCompleted(it.id)
-                                atLeastOneError = false
-                            } /*else { //default is true
-                                atLeastOneError = true
-                            }*/
-                        }
-                    }
-                }
-            }
-        }
-
-        return if (atLeastOneError) {
-            Response.Error(IOException("Some error happened during the upload process"))
-        } else {
-            Response.Success(Unit)
-        }
-    }
-
-    suspend fun purgeAllGeoTrackingDatapointAlreadyUploaded(): Response<Unit> {
-        GeoTrackingDatapointDao(herdrDb).deleteUploadedAll()
-        log.i { "Geolocation table was purged with success" }
-        return Response.Success(Unit)
     }
 
 //    @Update
